backend/requirements.txt
# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
alembic==1.12.1

# Redis
redis==5.0.1
hiredis==2.2.3

# Celery
celery==5.3.4

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# HTTP Requests
httpx==0.25.1
aiohttp==3.9.0
requests==2.31.0

# Web Scraping
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Utils
python-slugify==8.0.1
Pillow==10.1.0
pytz==2023.3
jdatetime==4.1.1

# Monitoring
sentry-sdk==1.38.0